% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomMatrixPackage.R
\name{create_random_matrix}
\alias{create_random_matrix}
\title{Create a random matrix}
\usage{
create_random_matrix(nrow, ncol)
}
\arguments{
\item{nrow}{Number of rows.}

\item{ncol}{Number of columns.}
}
\value{
A matrix with random numbers.
}
\description{
This function creates a random matrix of specified dimensions.
}
\examples{
create_random_matrix(3, 3)
}
