% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomMatrixPackage.R
\name{extract_random_number}
\alias{extract_random_number}
\title{Shuffle a matrix and extract a random number}
\usage{
extract_random_number(mat)
}
\arguments{
\item{mat}{A matrix to be shuffled.}
}
\value{
A random number from the shuffled matrix.
}
\description{
This function shuffles the elements of a matrix and returns a random number.
}
\examples{
mat <- create_random_matrix(3, 3)
extract_random_number(mat)
}
