% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomMatrixPackage.R
\name{permute_vector}
\alias{permute_vector}
\title{Permute a vector}
\usage{
permute_vector(vec)
}
\arguments{
\item{vec}{A vector to be permuted.}
}
\value{
A permuted vector.
}
\description{
This function generates a random permutation of a given vector.
}
\examples{
permute_vector(c(1, 2, 3, 4, 5))
}
