% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomMatrixPackage.R
\docType{package}
\name{RandomMatrixPackage}
\alias{RandomMatrixPackage-package}
\alias{RandomMatrixPackage}
\title{RandomMatrixPackage: A Package for Creating and Manipulating Random Matrices}
\description{
This package provides functions for creating and manipulating random matrices.
It includes functions to create random matrices, shuffle matrices, extract random numbers,
permute vectors, select random elements from matrices, and generate random walk matrices.
}
\section{Functions}{

\itemize{
  \item \code{\link{create_random_matrix}}: Create a random matrix.
  \item \code{\link{extract_random_number}}: Shuffle a matrix and extract a random number.
  \item \code{\link{permute_vector}}: Permute a vector.
  \item \code{\link{select_random_elements}}: Select random elements from a matrix.
  \item \code{\link{random_walk_matrix}}: Generate a random walk matrix.
}
}

\keyword{internal}
