% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomMatrixPackage.R
\name{select_random_elements}
\alias{select_random_elements}
\title{Select random elements from a matrix}
\usage{
select_random_elements(mat, num_elements)
}
\arguments{
\item{mat}{A matrix from which to select random elements.}

\item{num_elements}{Number of random elements to select.}
}
\value{
A vector of randomly selected elements from the matrix.
}
\description{
This function selects a specified number of random elements from a given matrix.
}
\examples{
mat <- create_random_matrix(3, 3)
select_random_elements(mat, 2)
}
